feat: added DTO, and DB Relations

+++ src/main/java/com/pos/hyper/repository/CustomerRepository.java
@@ -1,7 +1,7 @@
 package com.pos.hyper.repository;
 
-import com.pos.hyper.model.Customer;
+import com.pos.hyper.model.customer.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer,Long> {
+public interface CustomerRepository extends JpaRepository<Customer,Integer> {
 }
